+++ core/src/main/java/org/jdbi/v3/core/result/internal/ResultSetResultIterable.java
@@ -13,14 +13,21 @@
  */
 package org.jdbi.v3.core.result.internal;
 
+import java.lang.reflect.Type;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Collection;
+import java.util.List;
+import java.util.Set;
 import java.util.function.Supplier;
+import java.util.stream.Collector;
 
+import org.jdbi.v3.core.generic.GenericTypes;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.result.ResultIterable;
 import org.jdbi.v3.core.result.ResultIterator;
 import org.jdbi.v3.core.result.ResultSetException;
+import org.jdbi.v3.core.result.UnableToProduceResultException;
 import org.jdbi.v3.core.statement.StatementContext;
 
 public class ResultSetResultIterable<T> implements ResultIterable<T> {
@@ -48,5 +55,29 @@ public class ResultSetResultIterable<T> implements ResultIterable<T> {
         } catch (final SQLException e) {
             throw new ResultSetException("Unable to iterate result set", e, ctx);
         }
+    }
+
+    @Override
+    public <R extends Collection<? super T>> R collectInto(Type containerType) {
+
+        Type type = containerType;
+        if (containerType instanceof Class) {
+            // a raw class is a collection that needs to be parameterized with Object
+            type = GenericTypes.parameterizeClass((Class) containerType, Object.class);
+        }
+
+        Collector<? super T, ?, R> collector = (Collector<? super T, ?, R>) ctx.findCollectorFor(type)
+                .orElseThrow(() -> new UnableToProduceResultException("Could not find collector for " + containerType));
+        return collect(collector);
+    }
+
+    @Override
+    public List<T> collectIntoList() {
+        return collectInto(List.class);
+    }
+
+    @Override
+    public Set<T> collectIntoSet() {
+        return collectInto(Set.class);
     }
 }