+++ RELEASE_NOTES.md
@@ -13,6 +13,8 @@
   - StringTemplate engine: handle st4 errors rather than logging to stderr. Allow configuring missing attribute as a fatal error
   - StringTemplate 4.3.4
   - update spring framework dependency to 5.3.26 (address CVE-2023-20861)
+  - ResultIterable new set() and collectToMap() methods
+  - Helpers to make registering CollectorFactory easier
 
 # 3.37.1
   - fix deadlock in default Jdbi cache (#2274)