+++ RELEASE_NOTES.md
@@ -12,7 +12,7 @@
   - rewrite sqlobject and the generator to use the new extension framework, deprecate functionality that moved to the core
   - StringTemplate engine: handle st4 errors rather than logging to stderr. Allow configuring missing attribute as a fatal error
   - StringTemplate 4.3.4
-  - update spring framework dependency to 5.3.26 (address CVE-2023-20861)
+  - update spring framework dependency to 5.3.27 (address CVE-2023-20861, CVE-2023-20863)
   - ResultIterable new set() and collectToMap() methods
   - Helpers to make registering CollectorFactory easier
   - FreemarkerEngine encourages singleton use so caching works better