+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 3.38.0
+
   - `CaseStrategy` is now an official API (dropped @Beta) (#2309)
   - rewrite `AccessibleObjectStrategy` into an enum (#2310)
   - allow user defined functions for message rendering (#2311)